Fix ansible failure detection in clusterexec
The ansible transport provides a set of expected results for a playbook
execution.  If the actual result appears in the expected results, both
the actual result and set of gathered facts are returned.  Otherwise an
exception is raised.

clusterexec() was basing its failure detection on the returned result
value instead of whether or not an exception was raised.  This caused
it to misinterpret successful but non-zero results as failures.