bdd: Tests now run significantly faster.
Many of our BDD tests were creating new hosts before running tests as
they needed the host to exist. Since the test itself is not verifying
the creation of the host in those instances it didn't add any value to
the test while slowing things down. Tests that do not verify creation of
hosts now use a step noting that the host already exists which is
implemented by populating etcd directly.