compat: logging.config.dictConfig for < 2.7 added.
We are using logging.config.dictConfig to allow easy configuration of our loggers via configuration files. However, this functionality was added in 2.7. This change adds a basic implementation of logging.config.dictConfig to allow configuration on 2.6. It is important to note that our reimplementation is not nearly as exhaustive as the true 2.7+ implementation.