Launch investigator from a CherryPy plugin.
I noticed the investigator process was leaking in the BDD tests.
This was because the BDD framework was sending the server process
a SIGTERM, which terminated the main process immediately and left
the investigator child process running.

CherryPy has a signal handler plugin, so I tried enabling it.

That's when things got weird.

I suspect something about the way multiprocessing.Process is
implemented was causing cherrypy.engine.exit() to hang while
waiting for child threads to terminate on SIGTERM.

So I had the idea of launching the investigator process from a
CherryPy plugin to ensure it gets terminated.

Hence, CherryPyInvestigatorPlugin.

The other significant change is I gave the investigator its own
etcd.Client instance.  Since this runs in a subprocess, I don't
understand how it was interacting with CherryPy plugins before.
But however that worked my changes broke it, so I reverted back
to direct etcd.Client calls.