Use a thread pool instead of subprocesses
This replaces the InvestigatorPlugin with a JobsPlugin that provides
global access to a thread pool where async jobs (including clusterexec)
can be submitted.

At present the thread pool implementation is a bit of a hack from
Python's own standard library.  multiprocessing.pool.ThreadPool is
actually a multiprocessing.Pool that uses multiprocessing.dummy as
the Process object, which just wrappers the threading module.
Suffice it to say it offers little control over thread behavior.

Also notable is the investigator job no longer loops.