Added HostCredResource for RESTful credential requests.
HostCredResource allows for access to the ssh_priv_key and remote_user
which are part of the Host model but not provided on Host requests. This
has been specifically added so that client side applications which
require the ssh key can gain access and use it for their own
authentication. Example: client side ssh connections
(https://github.com/projectatomic/commctl/pull/17).