Add StoreHandlerManager.
This ties in to CPD-126 as an extra layer between CherryPyStorePlugin
and various StoreHandler instances.  The idea is StoreHandlerManager
can be cloned with an identical configuration and handed off to child
processes so they can read and write to the appropriate stores while
avoiding the forked CherryPy bus.

A couple small deviations from CPD-126:

1) The "register_store_handler" method takes a type object and a
   configuration object instead of a StoreHandler instance.  The
   type object should be derived from the StoreHandlerBase class.
   StoreHandlerManager will create instances of this type lazily
   as needed, and these instances are NOT duplicated in cloned
   StoreHandlerManagers.

2) Consequently, StoreHandler instances can go ahead and connect
   to their stores immediately.  So the "_get_connection" method
   is no longer needed.

Note the API still takes etcd key paths instead of model instances.
This is temporary so we can implement and test CPD-126 incrementally.

[1] https://github.com/projectatomic/commissaire/wiki/cpd-126