Add --register-store-handler command-line option.
Per CPD-126 [1].  This can also be specified as a list of JSON objects
in commissaire.conf.  The content of each store handler definition is
still mostly undocumented, but tends to get passed directly as keyword
arguments to some client object.

This will eventually replace the --etcd-uri, --kube-uri, and similar
command-line options but we're not quite there yet.

Also, StoreHandlerManager is now using real model instances to look up
store handlers, instead of the "bogus model" kludge.

[1] https://github.com/projectatomic/commissaire/wiki/cpd-126