Remove etcd.Client usage in main().
Require all configuration parameters to be locally defined and
disregard those in etcd.  Kubernetes-specific parameters have
to be defined in a 'register_store_handler' command-line option
or (preferrably) JSON configuration section.

This is to help reign in added complexity from user defined store
handlers.  Remains to be seen whether this is a temporary feature
removal.